+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - DAML language support (`.daml`)
 - New mutation operators:
-  - `CPS` (Controller Party Swap) and `CPR` (Controller Party Removal) for DAML
+  - `CPS` (Controller Party Swap), `CPR` (Controller Party Removal), and `SPS` (Signatory Party Swap) for DAML
 
 ## 3.1.0 - 2026-04-20